# fix FLOP/s calculation for OpenMP